Fix AutoMinorVersionUpgrade forcing error on update (#52)
Issue #, if available: [2723](https://github.com/aws-controllers-k8s/community/issues/2723)

Description of changes:
To avoid inclusion of AutoMinorVersionUpgrade in ModifyDBInstance payload causing API error for unrelated spec field changes omit the AutoMinorVersionUpgrade if it is not in the delta. 

- Omit AutoMinorVersionUpgrade if it is not included in the delta
- Add .vscode to .gitignore
- Rebuild controller with latest code-gen

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.